Modify TC, do refactoring
