+++ SSD/command_parse.cpp
@@ -1,0 +1,55 @@
+#include <iostream>
+#include <io.h>
+#include <iomanip>
+#include "command_parse.h"
+#include <string>
+
+using namespace std;
+
+void SSDCommand::WriteInvalidLog() {
+	FileObj->writeInvalidLog();
+}
+
+bool SSDCommand::checkInvalidity(const char& CMD, string LBAstring, char* data) {
+	if (!((CMD == 'W') || (CMD == 'w') || (CMD == 'R') || (CMD == 'r'))) {
+		WriteInvalidLog();
+		return false;
+	}
+
+	size_t pos;
+	uint32_t value = std::stoi(LBAstring, &pos, 10);
+	if (pos != LBAstring.length()) {
+		WriteInvalidLog();
+		return false;
+	}
+	if (value < 0 || value > 99) {
+		WriteInvalidLog();
+		return false;
+	}
+	if ((CMD == 'W') || (CMD == 'w')) {
+		std::string firstTwo(data, 2);
+		if (strlen(data) != 10 || firstTwo != "0x") {
+			WriteInvalidLog();
+			return false;
+		}
+		value = std::stoul(data, &pos, 16);
+		if (pos != strlen(data)) {
+			WriteInvalidLog();
+			return false;
+		}
+	}
+	return true;
+}
+
+void SSDCommand::parseArg(char CMD, string LBAstring, char* data) {
+	if (checkInvalidity(CMD, LBAstring, data) != true) return;
+
+	if ((CMD == 'W') || (CMD == 'w')) {
+		strcpy_s(input_data, data);
+		bIsWrite = true;
+	}
+	else if ((CMD == 'R') || (CMD == 'r')) {
+		bIsWrite = false;
+	}
+	LBA = std::stoi(LBAstring);	
+}
